+++ src/app/(route)/news/_utils/onHandleToTop.ts
@@ -1,0 +1,10 @@
+export const onHandleToTop = (commentBarRef: React.RefObject<HTMLDivElement>) => {
+  if (commentBarRef.current) {
+    const navBarHeight = 130;
+    const y =
+      commentBarRef.current.getBoundingClientRect().top +
+      window.scrollY -
+      navBarHeight;
+    window.scrollTo({ top: y, behavior: "smooth" });
+  }
+};
